requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pillow==10.1.0
pydantic==2.5.2
motor==3.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
